Add specific enum type resolution (#27)
* config-allowlist.go: Update blocked header imports

* intermediate.go: Rename enum methods

* main.go: Update enum method invocation

* statetracker.go: Add specific enum type resolution

* emitcabi.go: Use enum structures

* emitc.go: Use enum structures

* README: Update enum information

* Update wrappers

+++ include/libqarraydata.h
@@ -56,7 +56,7 @@ int64_t q_arraydata_detach_capacity(void* self, int64_t newSize);
 /// @param objectSize int64_t
 /// @param newCapacity int64_t
 /// @param option enum QArrayData__AllocationOption
-libqt_pair /* tuple of QArrayData* and void* */ q_arraydata_reallocate_unaligned(void* data, void* dataPointer, int64_t objectSize, int64_t newCapacity, int64_t option);
+libqt_pair /* tuple of QArrayData* and void* */ q_arraydata_reallocate_unaligned(void* data, void* dataPointer, int64_t objectSize, int64_t newCapacity, int32_t option);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qarraydata.html#deallocate)
 ///
